Move filters to package, to make them more testable
Testing techniques like this one https://www.evernote.com/shard/s209/sh/e2d21b27-2f4a-409a-891b-400e71a3236a/4b2dda3166be76a67bc070bffd49f5d3 require the tester to pass the filter class to a Mock annotation, which does not get resolved correctly when the class is not in a package.
