+++ grails-app/conf/com/netflix/asgard/AuthorizationFilters.groovy
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2013 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.netflix.asgard
+
+import com.netflix.asgard.plugin.AuthorizationProvider
+
+/**
+ * Invokes Shiro's access control if any of the configured {@link AuthorizationProvider} objects indicates the current
+ * subject does not have access to the requested endpoint.
+ */
+class AuthorizationFilters {
+
+    def pluginService
+
+    def filters = {
+        all(controller: '*', action: '*') {
+            before = {
+                Collection<AuthorizationProvider> authorizationProviders = pluginService.authorizationProviders
+                if (authorizationProviders.any { !it.isAuthorized(request, controllerName, actionName) }) {
+                    return accessControl()
+                }
+                true
+            }
+        }
+    }
+}
