+++ grails-app/conf/com/netflix/asgard/RegionFilters.groovy
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import com.netflix.asgard.Region
+package com.netflix.asgard
 
 class RegionFilters {
