+++ grails-app/conf/com/netflix/asgard/SettingsFilters.groovy
@@ -13,6 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package com.netflix.asgard
+
 import org.apache.shiro.SecurityUtils
 
 class SettingsFilters {
